Add option to compute ccd gray corrections at the sub-ccd scale.
This update works with the updated fgcm v1.9.9.1 which adds the ability to
compute ccd gray corrections with chebyshev polynomials, which are then
combined with the super star-flat parametrization.  A small half-pixel shift in
the chebyshev implementation in the stack and fgcm was found and tests were
added to check that everything works the same now.  The trigger of the final
fit cycle has been updated to simply setting "isFinalCycle=True", and this
functionality was added to the tests.  The hsc tests were updated to use the
new sub-ccd gray correction and comparison values were updated accordingly
(changes at the <0.1 mmag level).