Add ability to factor out wcs jacobian distortions from photom. calibration.
With new ProductBoundedField and PixelAreaBoundedField in afw we can properly
decompose the photometric calibration into the constituent parts.  The jacobian
computations are done using camera geometry because of performance issues (this
is much, much faster), and tests show that this is more reliable, and
consistent at the 1 mmag level with the predictions from the SkyWcs. Because of
the changes in the way the photometric calibration is computed (wcs jacobian +
illumination correction + gray), the test values have changed slightly and the
number of "good" stars in the calibration changed.  Finally, additional
configuration parameters for automatically selecting per-cycle photometric cuts
were added.