Add new "sedterms" for configurable magnitude to SED slope approximation.
Previously, fgcm required all bands be in order, and only adjacent colors would
be used.  The addition of NB filter support makes this difficult.  The new
"sedterms" have been added, and are now supported by fgcm, so that band
combinations are now much more flexible.  This makes it possible to use
chromatic terms in the tests (previously, 3 bands were required for chromatic
corrections, and now just two are required).  The test numbers have changed
slightly with the chromatic terms.