Add requireFiniteRaDec clause to source selector
We are now persisting source catalogs even in light of an astrometric
failure in single frame processing.  Such catalogs will have their sky
coordinate column (typically "coord_ra" and "coord_dec") values set to
NaN.  They are thus not suitable for inclusion in the fgcm calibration,
so this config default ensures they are omited from the selection of
"good" sources.