Rewrite lookupStaticCalibrations to avoid needing a special collection.
This is probably a bit slower than the previous version, because it
can use multiple smaller queries instead of a single larger one
(depending on dimensions).  But it's at most n_detectors or
n_physical_filters extra queries, and that's a small price to pay for
being able to get rid of the special "unbounded" collection.