Refactor connection building and add plot connections.
This does a refactor of the way the plot connections are built,
taking advantage of new functionality in pipe_base that obviates
the need for the class variable wrangling. It also changes dataset
names to be more easily sortable. Finally, it add a large number
of plot dataset types that can be serialized in fgcm.